Moves initial salesforce connection into separate module
